+++ CHANGELOG.md
@@ -39,7 +39,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Improvements
 
-- [#1524](https://github.com/babylonlabs-io/babylon/pull/1524) Remove `PropagateFPSlashingToConsumers` call from slashing logic at send finality signature
+- [#1525](https://github.com/babylonlabs-io/babylon/pull/1525) Remove `PropagateFPSlashingToConsumers` call from slashing logic at send
+finality signature and `SlashedBTCDelegation` structure.
 - [#1229](https://github.com/babylonlabs-io/babylon/pull/1229) Wire up EVM.
 
 ## v3.0.0-rc.0